blank (ts) installation
I had to create a new expo app. i hope this doesnt break anything
