+++ .gitignore
@@ -1,3 +1,5 @@
+client/.example
+
 # Node.js dependencies
 client/node_modules/
 client/.expo/
