+++ client/app/LoginScreen.tsx
@@ -1,0 +1,21 @@
+import { 
+    StyleSheet,
+    Text,
+    View,
+} from 'react-native';
+
+export default function LoginScreen() {
+    return (
+        <View style={styles.container}>
+            <Text>Login Screen</Text>
+        </View>
+    );
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+});
