+++ client/assets/splash-icon.png
